+++ production_ready_examples/iotics-connector-example-databypass/src/iotics/connector/example/databypass/main.py
@@ -2,15 +2,15 @@ from logging import config
 
 from constants import LOGGING_CONFIGURATION
 from data_processor import DataProcessor
-from follower_connector import FollowerConnector
+from databypass_connector import DataBypassConnector
 
 config.dictConfig(LOGGING_CONFIGURATION)
 
 
 def main():
-    data_processor = DataProcessor(use_db=True)
-    follower_connector = FollowerConnector(data_processor)
-    follower_connector.start()
+    data_processor = DataProcessor()
+    databypass_connector = DataBypassConnector(data_processor)
+    databypass_connector.start()
 
 
 if __name__ == "__main__":