+++ production_ready_examples/CHANGELOG.md
@@ -2,9 +2,17 @@
 
 All notable changes to these examples will be documented in this file.
 
+## 2024-08-05
+
+- Added **Historian Writer** and **Data Bypass** Connector examples.
+- Renamed **Follower Connector** to **Historian Writer** Connector.
+- Added `DBReader` and `DBWriter` classes to better manage read and write operations from/to the DB.
+- DB credentials moved from `constants.py` to `.env` file
+- Added `scope` parameter in `search_twins` method.
+
 ## 2024-05-28
 
-- Added Synthesiser Connector example.
+- Added **Synthesiser Connector** example.
 - Added additional methods to `data_processor.py` used by the Synthesiser Connector.
 - Added `use_db` parameter in `DataProcessor` constructor to handle the use of this class with Connectors that don't need the initialisation of `DBManager`.